pyproject.toml
[tool.poetry]
name = "tvsched"
version = "0.1.0"
description = ""
authors = ["astsu <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
asyncpg = "^0.25.0"
databases = {extras = ["postgresql"], version = "^0.5.3"}
passlib = "^1.7.4"
PyJWT = "^2.3.0"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
pytest = "^6.2.5"
pytest-asyncio = "^0.16.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
